Rework restrucutre to remove wrapper types
