--- src/builtins/native/timezone.rs
@@ -1,5 +1,0 @@
-use crate::tzdb::FsTzdbProvider;
-use std::sync::{LazyLock, Mutex};
-
-pub static TZ_PROVIDER: LazyLock<Mutex<FsTzdbProvider>> =
-    LazyLock::new(|| Mutex::new(FsTzdbProvider::default()));
